Move the CollectionTask.nonTaskUndo in Undo
The method's presence was not needed in the CollectionTask.
