+++ AnkiDroid/src/test/java/com/ichi2/libanki/sched/AbstractSchedTest.kt
@@ -18,7 +18,7 @@ package com.ichi2.libanki.sched
 import com.ichi2.anki.RobolectricTest
 import com.ichi2.anki.exception.ConfirmModSchemaException
 import com.ichi2.anki.servicelayer.Undo
-import com.ichi2.async.CollectionTask.Companion.nonTaskUndo
+import com.ichi2.anki.servicelayer.Undo.Companion.nonTaskUndo
 import com.ichi2.libanki.*
 import com.ichi2.libanki.utils.TimeManager.time
 import com.ichi2.testutils.AnkiAssert
